Makefile: remove object files from dependency graph
Using object files as dependencies of object files when these are not
the direct dependencies is misleading and can cause unexpected make
behaviour. Additionally, changes to inline functions in header files
will not cause recompilation of the object files which rely on them. The
correct approach is to include header files used by a source file in the
source file's object file's deps. This means that only interface changes
and inline changes cause recompilation of these objects.