Append and insert are the same on an empty buffer
Append and insert both insert a new character at index 0 on an empty buffer. 
Add check for the empty buffer before incrementing offset and Fixes #4.
The bug causes an append on the empty buffer to increment the offset 
to 1 instead of 0 and set the character at index 1, which is out of bounds
and segfaults.