Makefile patchset (#7)
* Makefile: remove uninstall target

An uninstall target is generally considered a danger, a few projects in
the past have been bitten by uninstall targets which did something wrong
and deleted someone's /usr directory. The general recommendation is to
leave uninstallation to package managers and users which avoids the
potential problems entirely.

* Makefile: remove install root check

It is an unusual, unnecessary and error prone thing to check if the user
is root while installing. Some users may want to install locally or
package the tool which means they won't need root since they will be
editing local files.

* Makefile: use $(RM) gmake variable

GNU Make (the make variant this makefile targets) has a "$(RM)"
variable. This patch changes the makefile to use this variable.

* Makefile: remove object files from dependency graph

Using object files as dependencies of object files when these are not
the direct dependencies is misleading and can cause unexpected make
behaviour. Additionally, changes to inline functions in header files
will not cause recompilation of the object files which rely on them. The
correct approach is to include header files used by a source file in the
source file's object file's deps. This means that only interface changes
and inline changes cause recompilation of these objects.

* Makefile: rename main.c to hx.c to simplify rules

Putting main in a file called <program>.c where the end result is
<program> can simplify GNU Makefiles. This means that all rules can be
made implicit.

Additionally, the implicit rule uses LDFLAGS for flags specific to
linking, this means LDFLAGS needs to be added to the makefile.

* Makefile: make "all" the first rule

Convention dictates that "all" should always be the first rule in a
makefile.

* Makefile: turn manpage gzip rule into pattern

Making the gzip compression rule used by the man page compression into a
pattern rule remove repetitive usage of filenames and makes the rule
more flexible for future use.

* Makefile: use -MMD -MP to generate deps automatically

Using the (gcc and clang supported) -MMD and -MP flags, files containing
makefile style dependency rules can be automatically generated from
source file during compilation. This means that the makefile can be
vastly simplified and the error prone process of maintaining a list of
header files on which each object file relies is no longer necessary.

These dependency files are only available after the first build but this
doesn't matter as they're only important during partial rebuilds which
can only happen after a project is initially built.

* Makefile: make install rule more flexible

By allowing for a DESTDIR, PREFIX and configuration of bindir and mandir
the makefile install rule can account for the vast majority of unusual
and usual installation configurations. This allows hx to be packaged by
distributions.

* Makefile: set HX_VERSION using git describe --tags

Where available git describe --tags will produce a unique and
incremental version number from past tags and hashes of changes since
the last tag. It's a good quick automatic solution for versioning. (But
can't completely replace manual version changes.)

* Makefile: move CPP flags to CPPFLAGS

CPPFLAGS is the variable used by gmake to specify C Pre-Processor flags.
It is more conventional to store CPPFLAGS separately from CFLAGS.

* Makefile: add a debug target

gmake allows setting variables when specific targets are made. This
allows a debug target to be created to move debug flags to be set by
this target. Passing -O multiple times means the last instance is
applied.

* Makefile: remove wildcards from clean

Using information we already have to remove specific files instead of
using wildcards removes the possibility of accidentally deleting files
which might have been misplaced.

* Makefile: general cleanup

This patch does some general cleanup to finish changes to the makefile.

This cleanup includes:
 - Splitting the setting of -D__BSD_VISIBLE to visually separate it from
   the rest of the assignment and make it clear that the variable is set
   only for BSD compatibility.
 - Adding spaces around some assignments as they had been inconsistently
   removed before (this improves readability).
 - Grouping .PHONY targets into one place to make the makefile look
   cleaner.
 - Removing unnecessary spacing around some targets to group them into
   related groups.
 - Moving the gzip pattern rule to the top of the file as a convention
   and to allow any future pattern rules to be grouped in one place.
 - Changing := $(shell ...) to != assignments in order to make the
   assignments visually simpler and easier to parse.
 - Removing certain comments which were obvious due to their context and
   therefore redundant (maintenance burden).
 - Using := for assignment of objects as this variable does not benefit
   from using any other kind of assignment.