implemented a start of :help command
This will display an on-line help, which is currently just a copy of the
command list in the README.md.

Also as a minor edit, changed a size to be unsigned, added an assert and
changed the Makefile a bit to allow assert to work when compiling with
`make debug`. I make this work by undefining NDEBUG, which looks a bit
hacky but it works, for now.