allow reading from the /proc/ filesystem (#10)
Since the procfs is not a real filesystem, stat() returns a zero content
length. It can be read until EOF is encountered however. Using fgetc()
we will try to read any file to which stat() returns a 0 length. For all
other files (where stat() returns a length > 0) we just fread() for
performance reasons.