rendering changed
Rendering of the ASCII values is different, and more like how the
filetype 'xxd' is rendered in vim. Meaning that the address is colored
purple, the hex line is green (my choice), and the ASCII equivalents
are drawn in such a way that printable characters are drawn in yellow
and non-printable in cyan.

This may be visually more pleasing (distinguishing
printable/non-printable is easier) but YMMV.